- Fixed a bug in the way 'offset' is calculated which was preventing files beyond a certain (large) size from opening.
git-svn-id: https://zoomwhee.mgh.harvard.edu/svn_devel/OpenMIMS/trunk@509 ea379df5-90a0-4b1b-ac45-2d7255ffb434

+++ src/com/nrims/data/Mims_Reader.java
@@ -93,7 +93,7 @@ public class Mims_Reader implements Opener {
         int bytesPerMass = pixelsPerImage * 2;
         short[] spixels = new short[pixelsPerImage];
 
-        long offset = dhdr.header_size + currentIndex * nMasses * bytesPerMass;
+        long offset = (long)dhdr.header_size + (long)currentIndex * (long)nMasses * (long)bytesPerMass;
         if (index > 0) {
             offset += index * bytesPerMass;
         }
